fix: Moving calssName to Util
