+++ templates/angular/ig-ts/custom-templates/editors-calculation-form/index.ts
@@ -20,7 +20,7 @@ class EditorsCalculationForm extends AngularTemplate {
 
 	public generateFiles(projectPath: string, name: string, ...options: any[]): Promise<boolean> {
 		const config = {
-			"$(ClassName)": this.className(name),
+			"$(ClassName)": Util.className(name),
 			"__name__": this.fileName(name),
 			"__path__": this.folderName(name)
 		};
