+++ templates/angular/ig-ts/grid/basic/index.ts
@@ -37,7 +37,7 @@ class GridTemplate extends AngularTemplate {
 
 	public generateFiles(projectPath: string, name: string, ...options: any[]): Promise<boolean> {
 		const config = {
-			"$(ClassName)": this.className(name),
+			"$(ClassName)": Util.className(name),
 			"__name__": this.fileName(name),
 			"__path__": this.folderName(name)
 		};
