+++ templates/angular/ig-ts/hierarchical-grid/hierarchical-grid-custom/index.ts
@@ -40,7 +40,7 @@ class HierarchicalGridCustomTemplate extends AngularTemplate {
 
 	public generateFiles(projectPath: string, name: string, ...options: any[]): Promise<boolean> {
 		const config = {
-			"$(ClassName)": this.className(name),
+			"$(ClassName)": Util.className(name),
 			"__name__": this.fileName(name),
 			"__path__": this.folderName(name)
 		};
