+++ templates/angular/ig-ts/hierarchical-grid/hierarchical-grid-export/index.ts
@@ -37,7 +37,7 @@ class HierarchicalGridExportTemplate extends AngularTemplate {
 
 	public generateFiles(projectPath: string, name: string, ...options: any[]): Promise<boolean> {
 		const config = {
-			"$(ClassName)": this.className(name),
+			"$(ClassName)": Util.className(name),
 			"__name__": this.fileName(name),
 			"__path__": this.folderName(name)
 		};
