+++ templates/react/es6/grid/grid-custom/index.ts
@@ -41,9 +41,9 @@ class GridCustomTemplate extends ReactTemplate {
 		const features = this.gridHelper.generateFeatures(this.userExtraConfiguration["features"], 5);
 
 		config["__path__"] =  this.folderName(name); //folder name allowed spaces, any casing
-		config["$(ClassName)"] = this.className(name); //first letter capital, no spaces,
+		config["$(ClassName)"] = Util.className(name); //first letter capital, no spaces,
 		config["$(widget)"] = "igGrid";
-		config["$(Control)"] = this.className("igGrid");
+		config["$(Control)"] = Util.className("igGrid");
 		config["$(igniteImports)"] = this.getImports();
 		config["$(name)"] = name; // this name should not have restrictions
 		config["$(description)"] = this.description;
