+++ templates/react/es6/hierarchical-grid/hierarchical-grid-custom/index.ts
@@ -40,9 +40,9 @@ class HierarchicalGridCustomTemplate extends ReactTemplate {
 		const features = this.gridHelper.generateFeatures(this.userExtraConfiguration["features"], 5);
 
 		config["__path__"] =  this.folderName(name); //folder name allowed spaces, any casing
-		config["$(ClassName)"] = this.className(name); //first letter capital, no spaces,
+		config["$(ClassName)"] = Util.className(name); //first letter capital, no spaces,
 		config["$(widget)"] = "igHierarchicalGrid";
-		config["$(Control)"] = this.className("igHierarchicalGrid");
+		config["$(Control)"] = Util.className("igHierarchicalGrid");
 		config["$(igniteImports)"] = this.getImports();
 		config["$(name)"] = name; // this name should not have restrictions
 		config["$(description)"] = this.description;
